Moved public API to api and api.builder packages
