+++ src/main/java/com/itquasar/multiverse/mail/api/EmailContact.java
@@ -1,4 +1,4 @@
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.api;
 
 import com.itquasar.multiverse.mail.exception.EmailException;
 import com.itquasar.multiverse.mail.util.Constants;
