+++ src/main/java/com/itquasar/multiverse/mail/api/Envelope.java
@@ -3,9 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.itquasar.multiverse.mail.message.envelope;
+package com.itquasar.multiverse.mail.api;
 
-import com.itquasar.multiverse.mail.EmailContact;
 import java.util.Date;
 import java.util.List;
