+++ src/main/java/com/itquasar/multiverse/mail/api/builder/EmailBuilderFactory.java
@@ -1,4 +1,4 @@
-package com.itquasar.multiverse.mail;
+package com.itquasar.multiverse.mail.api.builder;
 
 /**
  *
