+++ src/main/java/com/itquasar/multiverse/mail/java/JEmail.java
@@ -1,8 +1,8 @@
 package com.itquasar.multiverse.mail.java;
 
-import com.itquasar.multiverse.mail.message.content.Content;
+import com.itquasar.multiverse.mail.api.Content;
 import com.itquasar.multiverse.mail.message.ImmutableEmail;
-import com.itquasar.multiverse.mail.message.envelope.Envelope;
+import com.itquasar.multiverse.mail.api.Envelope;
 import javax.mail.Message;
 
 /**
