+++ src/main/java/com/itquasar/multiverse/mail/part/MimeTypes.java
@@ -1,6 +1,6 @@
 package com.itquasar.multiverse.mail.part;
 
-import com.itquasar.multiverse.mail.IMimeType;
+import com.itquasar.multiverse.mail.api.IMimeType;
 
 /**
  *
