+++ src/main/java/com/itquasar/multiverse/mail/util/ClientUtils.java
@@ -1,6 +1,6 @@
 package com.itquasar.multiverse.mail.util;
 
-import com.itquasar.multiverse.mail.EmailContact;
+import com.itquasar.multiverse.mail.api.EmailContact;
 import com.itquasar.multiverse.mail.part.MimeTypes;
 import com.itquasar.multiverse.mail.part.Part;
 import java.util.Arrays;
